ponderada-semana7/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
altair==5.3.0
ament-cmake-test==1.3.8
ament-copyright==0.12.10
ament-cppcheck==0.12.10
ament-cpplint==0.12.10
ament-flake8==0.12.10
ament-index-python==1.4.0
ament-lint==0.12.10
ament-lint-cmake==0.12.10
ament-package==0.14.0
ament-pep257==0.12.10
ament-uncrustify==0.12.10
ament-xmllint==0.12.10
angles==1.15.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
attrs==23.2.0
Babel==2.8.0
beautifulsoup4==4.10.0
beniget==0.4.1
blessed==1.20.0
blinker==1.4
bond==3.0.2
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.1
cachetools==5.3.3
caer==2.0.8
cartographer-ros-msgs==2.0.9002
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.16.1
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.1
control-msgs==4.4.0
controller==2023.1.2
controller-manager==2.40.0
controller-manager-msgs==2.40.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dwb-msgs==1.1.14
editor==1.6.6
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.3
gitdb==4.0.11
GitPython==3.1.43
gpg==1.16.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
inquirer==3.2.4
interactive-markers==2.3.2
jeepney==0.7.1
Jinja2==3.0.3
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.5
launch-ros==0.19.7
launch-testing==1.0.5
launch-testing-ros==0.19.7
launch-xml==1.0.5
launch-yaml==1.0.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.3
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mdurl==0.1.2
message-filters==4.3.3
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
mypy==1.10.0
mypy-extensions==1.0.0
natsort==8.0.2
nav-2d-msgs==1.1.14
nav-msgs==4.2.3
nav2-common==1.1.14
nav2-msgs==1.1.14
nav2-simple-commander==1.0.0
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
osrf-pycommon==2.0.2
packaging==21.3
pandas==2.2.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==16.1.0
pycairo==1.20.1
pycodestyle==2.8.0
pycollada==0.6
pycups==2.0.1
pydeck==0.9.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.18.0
PyGObject==3.42.1
pygraphviz==1.7
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.12
rcutils==5.1.5
readchar==4.1.0
referencing==0.35.1
reportlab==3.6.8
requests==2.32.2
resource-retriever==3.1.1
rich==13.7.1
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.9
ros2bag==0.15.9
ros2cli==0.18.9
ros2component==0.18.9
ros2controlcli==2.40.0
ros2doctor==0.18.9
ros2interface==0.18.9
ros2launch==0.19.7
ros2lifecycle==0.18.9
ros2multicast==0.18.9
ros2node==0.18.9
ros2param==0.18.9
ros2pkg==0.18.9
ros2run==0.18.9
ros2service==0.18.9
ros2topic==0.18.9
rosbag2-interfaces==0.15.9
rosbag2-py==0.15.9
rosdep==0.24.0
rosdep-modules==0.24.0
rosdistro-modules==0.9.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rospkg-modules==1.5.1
rpds-py==0.18.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.3
rqt-graph==1.3.0
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
runs==1.2.2
scipy==1.8.0
scripts==3.7.0
SecretStorage==3.3.1
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
shape-msgs==4.2.3
shellingham==1.5.4
six==1.16.0
slam-toolbox==2.6.8
smclib==3.0.2
smmap==5.0.1
snowballstemmer==2.2.0
soupsieve==2.3.1
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
streamlit==1.35.0
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.4.0
tenacity==8.3.0
tf2-geometry-msgs==0.25.6
tf2-kdl==0.25.6
tf2-msgs==0.25.6
tf2-py==0.25.6
tf2-ros-py==0.25.6
tf2-tools==0.25.6
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
topic-monitor==0.20.3
tornado==6.4
trajectory-msgs==4.2.3
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.4.2
typeguard==2.2.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
vehicle==2023.1.2
vision-msgs==4.1.1
visualization-msgs==4.2.3
wadllib==1.3.6
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
webots-ros2==2023.1.2
webots-ros2-driver==2023.1.2
webots-ros2-epuck==2023.1.2
webots-ros2-importer==2023.1.2
webots-ros2-mavic==2023.1.2
webots-ros2-msgs==2023.1.2
webots-ros2-tesla==2023.1.2
webots-ros2-tiago==2023.1.2
webots-ros2-turtlebot==2023.1.2
webots-ros2-universal-robot==2023.1.2
wxPython==4.0.7
xacro==2.0.8
xdg==5
xkit==0.0.0
xmod==1.8.1
zipp==1.0.0
